services/convertsvc/requirements.txt
fastapi
uvicorn
python-magic
python-multipart
pillow
